Init new tracer contract in Simulation
