+++ openfisca_core/tracers.py
@@ -249,3 +249,26 @@ class TracingParameterNodeAtInstant(object):
         if isinstance(child, (np.ndarray,) + ALLOWED_PARAM_TYPES):
             self.tracer.record_calculation_parameter_access(name, period, child)
         return child
+
+
+class SimpleTracer:
+
+    def __init__(self):
+        self._stack = []
+
+    @property
+    def stack(self):
+        return self._stack
+
+
+    @stack.setter
+    def stack(self, stack):
+        self._stack = stack
+
+
+    def enter_calculation(self, variable, period):
+        self.stack.append({'name': variable, 'period': period})
+
+
+    def exit_calculation(self):
+        self.stack.pop()
